Code testing cleanup.
